Accelerometers now outputs a double value instead of long

+++ Accelerometer.h
@@ -16,19 +16,19 @@ public:
 	 * Get accelerometer X value
 	 * @return X value
 	 */
-	virtual long getXAccel() = 0;
+	virtual double getXAccel() = 0;
 
 	/**
 	 * Get accelerometer Y value
 	 * @return Y value
 	 */
-	virtual long getYAccel() = 0;
+	virtual double getYAccel() = 0;
 
 	/**
 	 * Get accelerometer Z value
 	 * @return Z value
 	 */
-	virtual long getZAccel() = 0;
+	virtual double getZAccel() = 0;
 
 	/**
 	 * Get accelerometer name
